Move idle handling.
